Revert "Rename files to have internal- nomenclature"
This reverts commit 7911689f296f7e630ac6670334dd677d9db3fb1c.
